added ddata and spmd modes
